Complete T26 Core Detection Engine Integration and Windows Installer Fixes
- Integrate DetectionAlgorithmFactory into DuplicateDetector
  * Remove HashCalculator dependencies from duplicate_detector.h/.cpp
  * Unify async/sync detection workflows with algorithm-based signatures
  * Update DuplicateGroup to include algorithmUsed field

- Fix Windows NSIS Installer Creation
  * Resolve infinite loop issues in Qt deployment during CPack
  * Implement proper post-build Qt deployment with windeployqt
  * Include all Qt DLLs and plugin directories (platforms, iconengines, etc.)
  * Fix Qt platform plugin initialization errors

- Configure Cross-Platform Installers
  * Windows: NSIS generator for .exe installers
  * macOS: DragNDrop generator for .dmg disk images
  * Linux: DEB/RPM/TGZ generators for native packages

- Update TOMORROW_TASKS.md with T26 completion status
- Add GPU_IMPLEMENTATION_PLAN.md for future GPU acceleration

Phase 2 Core Integration: COMPLETE
Phase 3 Cross-Platform Installers: READY

Windows installer now includes all Qt dependencies and plugins,
enabling reliable distribution and testing of advanced detection algorithms.

+++ src/core/selection_history_manager.cpp
@@ -118,11 +118,11 @@ int SelectionHistoryManager::getMaxHistorySize() const {
 }
 
 int SelectionHistoryManager::getUndoStackSize() const {
-    return m_undoStack.size();
+    return static_cast<int>(m_undoStack.size());
 }
 
 int SelectionHistoryManager::getRedoStackSize() const {
-    return m_redoStack.size();
+    return static_cast<int>(m_redoStack.size());
 }
 
 void SelectionHistoryManager::enforceHistoryLimit() {
